Split microbatch_config into config, tensor ops and pipeline timing
microbatch_config.py had grown to ~1000 lines spanning four concerns.
It now keeps only the switches and logging helpers (333 lines); the
batch split/merge tensor ops move to microbatch_ops.py and the timing
tracker + AsyncOutputBuffer to pipeline_timing.py. Existing import
sites keep working through explicit re-exports.
